Estimate Difference in Survival or Cumulative Incidence and Confidence Interval

This function estimates the unadjusted difference in survival at a given time point based on the difference between per-group Kaplan-Meier estimates.

Usage

survdiff_ci(
  formula,
  data,
  time,
  estimand = c("survival", "cuminc"),
  conf.level = 0.95
)

Arguments

formula

Formula of a survival object using Surv of the form, Surv(time, event) ~ group. The exposure variable (here, group) must be categorical with at least 2 categories.

data

Data set.

time

Time point to estimate survival difference at.

estimand

Optional. Estimate difference in survival ("survival") or cumulative incidence ("cuminc")? This parameter affects the sign of the differences. Defaults to "survival".

conf.level

Optional. Confidence level. Defaults to 0.95.

Value

Tibble in tidy format:

term Name of the exposure stratum.

estimate Difference in survival at time.

std.error Large-sample standard error of the difference in survival functions (see References). For each survival function, Greenwood standard errors with log transformation are used, the default of the survival package/survfit).

statistic z statistic.

p.value

conf.low Lower confidence limit

conf.high Upper confidence limit

References

Com-Nougue C, Rodary C, Patte C. How to establish equivalence when data are censored: a randomized trial of treatments for B non-Hodgkin lymphoma. Stat Med 1993;12:1353–64. https://doi.org/10.1002/sim.4780121407.

Altman DG, Andersen PK. Calculating the number needed to treat for trials where the outcome is time to an event. BMJ 1999;319:1492–5. https://doi.org/10.1136/bmj.319.7223.14929.

Examples

# Load 'cancer' dataset from survival package (Used in all examples)
data(cancer, package = "survival")

cancer <- cancer %>%
  dplyr::mutate(
    sex = factor(
      sex,
      levels = 1:2,
      labels = c("Male", "Female")),
      status = status - 1)

survdiff_ci(
  formula = survival::Surv(time = time, event = status) ~ sex,
  data = cancer,
  time = 365.25)
#> # A tibble: 1 × 7
#>   term   estimate std.error statistic p.value conf.low conf.high
#>   <chr>     <dbl>     <dbl>     <dbl>   <dbl>    <dbl>     <dbl>
#> 1 Female    0.190    0.0739      2.58 0.00497   0.0456     0.335
# Females have 19 percentage points higher one-year survival than males
# (95% CI, 5 to 34 percentage points).